% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_flux.R
\name{n2o}
\alias{n2o}
\alias{tidy_licor}
\title{tidy_licor}
\usage{
tidy_licor(file_path, gas, analyzer = "licor")
}
\arguments{
\item{file_path}{Directory of file.}

\item{gas}{Choose between CO2/CH4 or N2O LI-COR Trace Gas Analyzer, which is "ch4" and "n2o", respectively.}

\item{analyzer}{The brand of the analyzer which the data was downloaded from.}
}
\description{
Tidy the data downloaded from GHG Analyzer.

Created for troubleshooting.
}
\examples{
\dontrun{
ch4 <- tidy_licor("ch4.xlsx", "ch4")
}
}
