% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_do.R
\name{calculate_do}
\alias{calculate_do}
\title{calculate_do}
\usage{
calculate_do(df)
}
\arguments{
\item{df}{Merged dataframe produced by process_hobo(), process_weather() and process_info() functions.}
}
\value{
A dataframe.
}
\description{
Calculate the Net Ecosystem Production,
Gross Primary Production and Ecosystem respiration based on the change in dissolved oxygen concentration.
}
\examples{
data(hobo)
calculate_do(hobo)
}
