% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_do.R
\name{combine_hobo}
\alias{combine_hobo}
\title{combine_hobo}
\usage{
combine_hobo(file_path, file_prefix = "no.")
}
\arguments{
\item{file_path}{Directory of file.}

\item{file_prefix}{Specify a prefix that the file names must start with in the file directory.}
}
\value{
A dataframe.
}
\description{
Apply the process_hobo function to multiple file at once.
}
\examples{
\dontrun{
df <- combine_hobo(file_path = "data/", file_prefix = "hobo_")
}
}
