% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_do.R
\name{combine_weather}
\alias{combine_weather}
\title{combine_weather}
\usage{
combine_weather(file_path, start_date, end_date, zone)
}
\arguments{
\item{file_path}{Directory of file, also need to include the character in the file name before the date.}

\item{start_date}{Start date of the file of daily weather data in yyyy-mm-dd format.}

\item{end_date}{End date of the file of daily weather data in yyyy-mm-dd format.}

\item{zone}{Code for the region of the weather station.}
}
\description{
Apply the process_weather function to multiple file at once.
}
\examples{
\dontrun{
df <- combine_weather(file_path = "weather/1B3D5F-", start_date = "2024-01-01",
end_date = "2024-01-07", zone = "site_A")
}
}
