% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHG_LICOR.R
\name{convert_time}
\alias{convert_time}
\title{convert_time}
\usage{
convert_time(data, day = 0, hr = 0, min = 0, sec = 0)
}
\arguments{
\item{data}{Data from the LI-COR Trace Gas Analyzer that had been processed by tidy_licor().}

\item{day}{Day(s) to add or subtract.}

\item{hr}{Hour(s) to add or subtract.}

\item{min}{Minute(s) to add or subtract.}

\item{sec}{Second(s) to add or subtract.}
}
\description{
Convert the time of the LI-COR Trace Gas Analyzer to match the time in real life.
}
\examples{
\dontrun{
ch4 <- convert_time(ch4, day = 1, min = 15)
}
}
