% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2o.R
\name{n2o}
\alias{n2o}
\title{Processed data from N2O LI-COR Trace Gas Analyzer.
A dataset containing 567 N2O concentrations changed over time.}
\format{
A data.frame with 567 rows and 4 variables:
\itemize{
  \item DATE: Date in character format.
  \item TIME: Time in character format.
  \item N2O: Concentrations of nitrous oxide (N2O), in ppb.
  \item date_time: Date and time in POSIXct format.
}
}
\source{
own data.
}
\description{
Processed data from N2O LI-COR Trace Gas Analyzer.
A dataset containing 567 N2O concentrations changed over time.
}
