% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_do.R
\name{plot_hobo}
\alias{plot_hobo}
\title{plot_hobo}
\usage{
plot_hobo(df)
}
\arguments{
\item{df}{Dataframe produced by process_hobo() function.}
}
\value{
A plot generated by ggplot2.
}
\description{
Plot the dissolved oxygen concentration over time series grouped by different data loggers to observe the variations.
}
\examples{
data(hobo)
plot_hobo(hobo)
}
