% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_do.R
\name{process_info}
\alias{process_info}
\title{process_info}
\usage{
process_info(file_path)
}
\arguments{
\item{file_path}{Directory of file.}
}
\value{
A dataframe.
}
\description{
Import and process the necessary information,
including the sunrise and sunset times of the day,
the date and time range of the deployment,
and the code for the data logger.
}
\examples{
info_data_path <- system.file("extdata", "info.xlsx", package = "aelab")
df <- process_info(info_data_path)
}
