% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_do.R
\name{process_weather}
\alias{process_weather}
\title{convert_time}
\usage{
process_weather(file_path, date, zone)
}
\arguments{
\item{file_path}{Directory of file.}

\item{date}{Date of the daily weather data in yyyy-mm-dd format.}

\item{zone}{Code for the region of the weather station.}
}
\description{
Tidy the daily weather data downloaded from weather station in Taiwan.
}
\examples{
\dontrun{
df <- process_weather("weather/day_one.csv", "2024-01-01", "site_A")
}
}
