% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scale_fill_aelab_c}
\alias{scale_fill_aelab_c}
\title{scale_fill_aelab_c}
\usage{
scale_fill_aelab_c(name, direction = 1)
}
\arguments{
\item{name}{Palette name passed to \code{\link{aelab_palettes}}.}

\item{direction}{\code{1} (default) for normal order; \code{-1} to reverse.}
}
\value{
A \code{ggplot2} scale.
}
\description{
Continuous ggplot2 fill scale using an aelab palette.
}
\examples{
\dontrun{
ggplot2::ggplot(mtcars, ggplot2::aes(factor(cyl), fill = mpg)) +
  ggplot2::geom_col() + scale_fill_aelab_c("ghg")
}
}
