% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHG_LICOR.R
\name{tidy_licor}
\alias{tidy_licor}
\title{tidy_licor}
\usage{
tidy_licor(file_path, gas)
}
\arguments{
\item{file_path}{Directory of file.}

\item{gas}{Choose between CO2/CH4 or N2O LI-COR Trace Gas Analyzer, which is "ch4" and "n2o", respectively.}
}
\description{
Tidy the data downloaded from LI-COR Trace Gas Analyzer.
}
\examples{
\dontrun{
ch4 <- tidy_licor("ch4.xlsx", "ch4")
}
}
