% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20231207_myRcode_v1.R
\name{calculate_regression}
\alias{calculate_regression}
\title{calculate_regression}
\usage{
calculate_regression(
  data,
  date,
  time,
  ghg,
  reference_time,
  duration_minutes = 7,
  num_rows = 300
)
}
\arguments{
\item{data}{File downloaded containing data of GHG concentration.}

\item{date}{Column name of the file containing date information (e.g., "DATE").}

\item{time}{Column name of the file containing time information (e.g., "TIME").}

\item{ghg}{Column name of the file containing data on GHG concentration (e.g., "CH4", "N2O").}

\item{reference_time}{The start time at which the measurement took place.}

\item{duration_minutes}{The duration(minutes) of the measurement, default to 7.}

\item{num_rows}{The number of rows used to perform the regression, default to 300.}
}
\description{
Calculate the slope of greenhouse gas (GHG) concentration change over time.
}
