% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_do.R
\name{combine_hobo}
\alias{combine_hobo}
\title{combine_hobo}
\usage{
combine_hobo(file_path, file_prefix = "no.")
}
\arguments{
\item{file_path}{Directory of the folder containing the files.}

\item{file_prefix}{The prefix before the code for the data logger, defaults to "no."}
}
\value{
A dataframe.
}
\description{
Tidy multiple data retrieved from HOBO U26 Dissolved Oxygen Data Logger.
}
\examples{
hobo_data_path <- system.file("extdata", package = "aelab")
df <- combine_hobo(hobo_data_path, file_prefix = "ex_ho")
}
