% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_do.R
\name{combine_weather}
\alias{combine_weather}
\title{combine_weather}
\usage{
combine_weather(file_path, start_date, end_date, zone)
}
\arguments{
\item{file_path}{Directory of folder containing the files (including the character in the file name that precedes the date).}

\item{start_date}{Date of the daily weather data in yyyy-mm-dd format.}

\item{end_date}{Date of the daily weather data in yyyy-mm-dd format.}

\item{zone}{Code for the region of the weather station.}
}
\value{
A dataframe.
}
\description{
Tidy multiple daily weather data downloaded from weather station in Taiwan.
}
\examples{
weather_data_path <- system.file("extdata", package = "aelab")
modified_data_path <- paste0(weather_data_path, "/ex_")
df <- combine_weather(modified_data_path,
start_date = "2024-01-01",
end_date = "2024-01-02", "site_A")
}
