% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_flux.R
\name{convert_time}
\alias{convert_time}
\title{convert_time}
\usage{
convert_time(data, day = 0, hr = 0, min = 0, sec = 0)
}
\arguments{
\item{data}{Data from the LI-COR Trace Gas Analyzer that had been processed by tidy_licor().}

\item{day}{Day(s) to add or subtract.}

\item{hr}{Hour(s) to add or subtract.}

\item{min}{Minute(s) to add or subtract.}

\item{sec}{Second(s) to add or subtract.}
}
\value{
The input data with a new column in POSIXct format converted based on the input value.
}
\description{
Convert the time of the LI-COR Trace Gas Analyzer to match the time in real life.
}
\examples{
data(n2o)
converted_n2o <- convert_time(n2o, min = -10, sec = 5)
}
