% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20231207_myRcode_v1.R
\name{convert_time}
\alias{convert_time}
\title{convert time}
\usage{
convert_time(datetime_value, day = 0, hr = 0, min = 0, sec = 0)
}
\arguments{
\item{datetime_value}{Recorded time in the unit of year/month/day hour:minutes:seconds.}

\item{day}{The day(s) difference between real time and the LI-COR Trace Gas Analyzer.}

\item{hr}{The hour(s) difference between real time and the LI-COR Trace Gas Analyzer.}

\item{min}{The minute(s) difference between real time and the LI-COR Trace Gas Analyzer.}

\item{sec}{The second(s) difference between real time and the LI-COR Trace Gas Analyzer.}
}
\description{
Convert the recorded time to match the time of the LI-COR Trace Gas Analyzer.
}
