% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_do.R
\name{process_hobo}
\alias{process_hobo}
\title{process_hobo}
\usage{
process_hobo(file_path, no_hobo)
}
\arguments{
\item{file_path}{Directory of file.}

\item{no_hobo}{The code for the data logger.}
}
\value{
A dataframe.
}
\description{
Tidy the data retrieved from HOBO U26 Dissolved Oxygen Data Logger.
}
\examples{
hobo_data_path <- system.file("extdata", "ex_hobo.csv", package = "aelab")
df <- process_hobo(hobo_data_path, "code_for_logger")
}
