% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_do.R
\name{process_hobo}
\alias{process_hobo}
\title{process_hobo}
\usage{
process_hobo(file_path, no_hobo)
}
\arguments{
\item{file_path}{Directory of file.}

\item{no_hobo}{The code for the data logger.}
}
\description{
Tidy the data retrieved from HOBO U26 Dissolved Oxygen Data Logger.
}
\examples{
\dontrun{
df <- process_hobo("data/hobo.csv", "no.1")
}
}
