% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_do.R
\name{process_weather}
\alias{process_weather}
\title{convert_time}
\usage{
process_weather(file_path, date, zone)
}
\arguments{
\item{file_path}{Directory of file.}

\item{date}{Date of the daily weather data in yyyy-mm-dd format.}

\item{zone}{Code for the region of the weather station.}
}
\value{
A dataframe.
}
\description{
Tidy the daily weather data downloaded from weather station in Taiwan.
}
\examples{
weather_data_path <- system.file("extdata", "ex_weather.csv", package = "aelab")
df <- process_weather(weather_data_path, "2024-01-01", "site_A")
}
